Move Author DTOs up
