+++ src-tauri/libcalibre/src/api/authors.rs
@@ -4,7 +4,7 @@ use std::sync::Mutex;
 use diesel::prelude::*;
 use diesel::SelectableHelper;
 
-use crate::application::services::domain::author::dto::NewAuthorDto;
+use crate::dtos::author::NewAuthorDto;
 use crate::domain::author::entity::NewAuthor;
 use crate::Author;
