--- src-tauri/libcalibre/src/application/services/domain/author/mod.rs
@@ -1,1 +1,0 @@
-pub mod dto;
