+++ src-tauri/libcalibre/src/application/services/domain/mod.rs
@@ -1,3 +1,2 @@
-pub mod author;
 pub mod book;
 pub mod file;
