+++ src-tauri/libcalibre/src/dtos/mod.rs
@@ -1,0 +1,1 @@
+pub mod author;
